An Arsenal of 'R' Functions for Large-Scale Statistical Summaries

An Arsenal of 'R' functions for large-scale statistical summaries, which are streamlined to work within the latest reporting tools in 'R' and 'RStudio' and which use formulas and versatile summary statistics for summary tables and models.

Details

The package download, NEWS, and README are available on CRAN: https://cran.r-project.org/package=arsenal

Functions

Below are listed some of the most widely used functions available in arsenal:

tableby: Summary statistics of a set of independent variables by a categorical variable.

paired: Summary statistics of a set of independent variables paired across two timepoints.

modelsum: Fit models over each of a set of independent variables with a response variable.

freqlist: Approximate the output from SAS's PROC FREQ procedure when using the /list option of the TABLE statement.

comparedf: Compare two data.frames and report any differences between them, much like SAS's PROC COMPARE procedure.

write2word, write2html, write2pdf: Functions to output tables to a single Word, HTML, or PDF document.

write2: Functions to output tables to a single document. (Also the S3 backbone behind the write2* functions.)

keep.labels: Keep the 'label' attribute on an R object when subsetting.

formulize: A shortcut to generate one-, two-, or many-sided formulas.

mdy.Date and Date.mdy: Convert numeric dates for month, day, and year to Date object, and vice versa.

is.Date: Test if an object is a date.

%nin%: Test for "not in".

allNA and includeNA: some useful functions for dealing with NAs.

Data

mockstudy: Mock study data for examples.

Examples

library(arsenal)