formulize

A shortcut to generate one-, two-, or many-sided formulas from vectors of variable names.

formulize(y = "", x = "", ..., data = NULL, collapse = "+",
  collapse.y = collapse)

Arguments

y, x, ...
Character vectors, names, or calls to be collapsed (by "+") and put left-to-right in the formula. If data is supplied, these can also be numeric, denoting which column name to use. See examples.

data
An R object with non-null column names.

collapse
How should terms be collapsed? Default is addition.

collapse.y
How should the y-terms be collapsed? Default is addition. Also accepts the special string "list", which combines them into a multiple-left-hand-side formula, for use in other functions.

See also

Examples

## two-sided formula
formulize("y", c("x1", "x2", "x3"))

#> y ~ x1 + x2 + x3
#> <environment: 0x6e19db8>

## one-sided formula
formulize(x = c("x1", "x2", "x3"))

#> ~x1 + x2 + x3
#> <environment: 0x6e19db8>

## multi-sided formula
formulize("y", c("x1", "x2", "x3"), c("z1", "z2"), "w1")

#> y ~ x1 + x2 + x3 ~ z1 + z2 ~ w1
#> <environment: 0x6e19db8>

## can use numerics for column names
data(mockstudy)
formulize(y = 1, x = 2:4, data = mockstudy)

#> case ~ age + arm + sex
#> <environment: 0x6e19db8>

## mix and match
formulize(1, c("x1", "x2", "x3"), data = mockstudy)

#> case ~ x1 + x2 + x3
#> <environment: 0x6e19db8>

## get an interaction
formulize("y", c("x1*x2", "x3"))

#> y ~ x1 * x2 + x3
#> <environment: 0x6e19db8>

## get only interactions
formulize("y", c("x1", "x2", "x3"), collapse = "*")

#> y ~ x1 * x2 * x3
#> <environment: 0x6e19db8>

## no intercept
formulize("y", "x1 - 1")

#> y ~ x1 - 1
#> <environment: 0x6e19db8>

formulize("y", c("x1", "x2", "-1"))

#> y ~ x1 + x2 + -1
#> <environment: 0x6e19db8>

## LHS as a list to use in arsenal functions
formulize(c("y1", "y2", "y3"), c("x", "z"), collapse.y = "list")

#> list(y1, y2, y3) ~ x + z
#> <environment: 0x6e19db8>

## use in an lm
form <- formulize(2, 3:4, data = mockstudy)
summary(lm(form, data = mockstudy))

#> 
#> Call:
#> lm(formula = form, data = mockstudy)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -41.800  -7.568   0.892   8.432  29.124 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   60.1075     0.5966 100.744   <2e-16 ***
#> armF: FOLFOX   0.6927     0.7088   0.977   0.3286    
#> armG: IROX     0.1484     0.8118   0.183   0.8550    
#> sexFemale     -1.2319     0.6105  -2.018   0.0438 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 11.51 on 1495 degrees of freedom
#> Multiple R-squared:  0.003365,	Adjusted R-squared:  0.001365 
#> F-statistic: 1.683 on 3 and 1495 DF,  p-value: 0.1688
#> 

## using non-syntactic names or calls (like reformulate example)
formulize(as.name("+-"), c("`P/E`", "`% Growth`"))

#> `+-` ~ `P/E` + `% Growth`
#> <environment: 0x6e19db8>

f <- Surv(ft, case) ~ a + b
formulize(f[[2]], f[[3]])

#> Surv(ft, case) ~ a + b
#> <environment: 0x6e19db8>